scripts/requirements.txt
openai
youtube-transcript-api
python-dotenv
numpy
tqdm
scrapetube
requests
